# src/admin/products/admin-products.controller.ts
import {
  Body,
  Controller,
  Delete,
  Get,
  Param,
  Patch,
  Post,
  Query,
  UseGuards,
  UseInterceptors,
} from '@nestjs/common';
import { JwtAuthGuard } from '../../auth/jwt-auth.guard';
import { Role } from '../../generated/prisma/client';
import { Roles } from '../../auth/roles.decorator';
import { RolesGuard } from '../../auth/roles.guard';
import { AuditInterceptor } from '../audit/audit.interceptor';
import { AdminProductsQueryDto } from '../../products/dto/admin-products-query.dto';
import { CreateProductDto } from '../../products/dto/create-product.dto';
import { CreateVariantDto } from '../../products/dto/create-variant.dto';
import { UpdateProductDto } from '../../products/dto/update-product.dto';
import { UpdateVariantDto } from '../../products/dto/update-variant.dto';
import { ProductsService } from '../../products/products.service';
import { ApiTags } from '@nestjs/swagger';

@UseGuards(JwtAuthGuard, RolesGuard)
@Roles(Role.admin)
@UseInterceptors(AuditInterceptor)
@ApiTags('admin')
@Controller('admin/products')
export class AdminProductsController {
  constructor(private readonly productsService: ProductsService) {}

  @Get()
  findAll(@Query() query: AdminProductsQueryDto) {
    return this.productsService.findAll(query);
  }

  @Get(':id')
  findOne(@Param('id') id: string) {
    return this.productsService.findOne(id);
  }

  @Post()
  create(@Body() dto: CreateProductDto) {
    return this.productsService.create(dto);
  }

  @Patch(':id')
  update(@Param('id') id: string, @Body() dto: UpdateProductDto) {
    return this.productsService.update(id, dto);
  }

  @Delete(':id')
  remove(@Param('id') id: string) {
    return this.productsService.remove(id);
  }

  @Post(':id/variants')
  addVariant(@Param('id') id: string, @Body() dto: CreateVariantDto) {
    return this.productsService.addVariant(id, dto);
  }

  @Patch(':id/variants/:variantId')
  updateVariant(
    @Param('variantId') variantId: string,
    @Body() dto: UpdateVariantDto,
  ) {
    return this.productsService.updateVariant(variantId, dto);
  }

  @Delete(':id/variants/:variantId')
  removeVariant(@Param('variantId') variantId: string) {
    return this.productsService.removeVariant(variantId);
  }

  /** Regenera el SKU de la variante con el esquema canónico `<slug(producto)>-<slug(color)>`. */
  @Post(':id/variants/:variantId/regenerate-sku')
  regenerateVariantSku(@Param('variantId') variantId: string) {
    return this.productsService.regenerateVariantSku(variantId);
  }
}